Refactor: Move MetadataSet outside fccloudapi.php
